Use attohttpc instead of minreq =D
