+++ Cargo.toml
@@ -21,6 +21,6 @@ flate2 = { version = "1.0.27", features = ["default", "rust_backend"] }
 html2text = { version = "0.6.0", features = ["ansi_colours"] }
 serde = { version = "1.0.188", features = ["derive"] }
 toiletcli = { version = "0.9.3", features = ["colors", "flags"] }
+attohttpc = { version = "0.26.1", features = ["tls-native"] }
 serde_json = "1.0.106"
 tar = "0.4.40"
-minreq = { version = "2.10.0", features = ["https"] }
