+++ src/download.rs
@@ -1,11 +1,121 @@
+use std::fs::{File, create_dir_all, remove_dir_all, remove_file};
+use std::io::{Read, Write, BufReader, BufWriter};
+
+use attohttpc::get;
+
+use flate2::read::GzDecoder;
+use tar::Archive;
+
 use toiletcli::flags::*;
 use toiletcli::flags;
 
-use crate::docs::{deserialize_docs_json, download_docset_tar_gz, extract_docset_tar_gz};
+use crate::docs::{deserialize_docs_json, Docs};
 
 use crate::common::ResultS;
-use crate::common::{is_docs_json_exists, is_docset_downloaded, is_docset_in_docs, get_docset_path};
-use crate::common::{BOLD, GREEN, PROGRAM_NAME, RESET, YELLOW};
+use crate::common::{is_docs_json_exists, is_docset_downloaded, is_docset_in_docs, get_docset_path,
+                    get_program_directory};
+
+use crate::common::{DEFAULT_DOWNLOADS_LINK, DEFAULT_USER_AGENT, VERSION, BOLD, GREEN, PROGRAM_NAME, RESET, YELLOW};
+
+fn download_docset_tar_gz_with_progress(docset_name: &String, docs: &Vec<Docs>) -> Result<(), String> {
+    let user_agent = format!("{DEFAULT_USER_AGENT}/{VERSION}");
+
+    for entry in docs.iter() {
+        if docset_name == &entry.slug {
+            let docsets_path = get_program_directory()?.join("docsets");
+            let specific_docset_path = docsets_path.join(&docset_name);
+
+            if !specific_docset_path.exists() {
+                create_dir_all(&specific_docset_path)
+                    .map_err(|err| format!("Cannot create `{docset_name}` directory: {err}"))?;
+            }
+
+            let tar_gz_path = specific_docset_path
+                .join(docset_name)
+                .with_extension("tar.gz");
+
+            let file = File::create(&tar_gz_path)
+                .map_err(|err| format!("Could not create `{tar_gz_path:?}`: {err}"))?;
+
+            let download_link = format!("{DEFAULT_DOWNLOADS_LINK}/{docset_name}.tar.gz");
+
+            let response = get(&download_link)
+                .header_append("user-agent", &user_agent)
+                .send()
+                .map_err(|err| format!("Could not GET {download_link}: {err}"))?;
+
+            let content_length = response.headers()
+                .get("content-length")
+                .map_or("0", |header| header.to_str().unwrap_or("0"))
+                .parse::<usize>()
+                .unwrap_or(0);
+
+            let mut file_writer = BufWriter::new(file);
+            let mut response_reader = BufReader::new(response);
+
+            let mut buffer = [0; 1024 * 4];
+            let mut file_size = 0;
+
+            while let Ok(size) = response_reader.read(&mut buffer) {
+                if size == 0 {
+                    break;
+                }
+
+                file_writer.write(&buffer[..size])
+                    .map_err(|err| format!("Could not download file: {err}"))?;
+
+                file_size += size;
+
+                print!("\rDownloading `{docset_name}`: received {file_size} of {content_length} bytes...");
+            }
+            println!();
+
+            file_writer.flush()
+                .map_err(|err| format!("Could not flush buffer to file: {err}"))?;
+
+            if file_size != content_length {
+                let message = format!(
+                    "File size ({file_size}) is different than required size ({content_length}). \
+                     Please re-run this command :("
+                    );
+
+                remove_dir_all(&specific_docset_path)
+                    .map_err(|err| format!("Could not remove bad docset ({specific_docset_path:?}): {err}"))?;
+
+                return Err(message);
+            }
+        }
+    }
+
+    Ok(())
+}
+
+
+fn extract_docset_tar_gz(docset_name: &String) -> Result<(), String> {
+    let docset_path = get_docset_path(docset_name)?;
+
+    if !docset_path.exists() {
+        create_dir_all(&docset_path)
+            .map_err(|err| format!("Cannot create `{docset_name}` directory: {err}"))?;
+    }
+
+    let tar_gz_path = docset_path.join(docset_name).with_extension("tar.gz");
+
+    let tar_gz_file =
+        File::open(&tar_gz_path).map_err(|err| format!("Could not open {tar_gz_path:?}: {err}"))?;
+
+    let reader = BufReader::new(tar_gz_file);
+    let tar = GzDecoder::new(reader);
+    let mut archive = Archive::new(tar);
+
+    archive
+        .unpack(docset_path)
+        .map_err(|err| format!("Could not extract {tar_gz_path:?}: {err}"))?;
+
+    remove_file(&tar_gz_path).map_err(|err| format!("Could not remove {tar_gz_path:?}: {err}"))?;
+
+    Ok(())
+}
 
 fn show_download_help() -> ResultS {
     let help = format!(
@@ -60,8 +170,8 @@ where
                 continue;
             }
 
-            println!("Downloading `{docset}`...");
-            download_docset_tar_gz(docset, &docs)?;
+            print!("Downloading `{docset}`...");
+            download_docset_tar_gz_with_progress(docset, &docs)?;
 
             println!("Extracting `{docset}` to `{}`...", get_docset_path(docset)?.display());
             extract_docset_tar_gz(docset)?;
